requirements.txt
keras
tensorflow
opencv-python
numpy
pyserial
pyaer # needs libcaer; see https://github.com/duguyue100/pyaer

